requirements.txt
streamlit>=1.35.0
anthropic>=0.25.0
python-docx>=1.1.0
docx2pdf>=0.1.8
pywin32>=311
